Crate shallow_water

Modules

balinit
Re-initialises a flow with balanced fields obtained from the conditions delta_t=gamma_t=0 using data previously set up with a data generation routine. Assumes the previous data has delta = gamma = 0.

constants

nhswps

parameters

spectral
Module containing subroutines for spectral operations, inversion, etc

sta2dfft
This module performs FFTs in two directions on two dimensional arrays using the stafft library module to actually compute the FFTs. If FFTs in one direction only are required use the stafft module directly. The module can compute any combination of sine, cosine and full FFTs in each direction. Along with the usual forwards (physical -> Fourier space) and reverse (Fourier space -> physical) routines there are also routines for computing the first derivatives in either direction.

stafft
Fourier transform module. This is not a general purpose transform package but is designed to be quick for arrays of length 2^n. It will work if the array length is of the form 2^i * 3^j * 4^k * 5^l * 6^m (integer powers obviously).

swto3d

utils

vstrip

Macros

array2_from_file

array3_from_file